avcodec/cfdf_d5video: use int8_t instead of int

+++ libavcodec/cfdf_d5video.c
@@ -216,7 +216,7 @@ static void df_decode_step(CFDFD5VideoContext *c,
         if (B == 40)
             continue;                                /* skip row */
         if (B >= 44 && B <= 72) {                   /* vertical copy */
-            static const int voff[8] = {-4,-3,-2,-1,1,2,3,4};
+            static const int8_t voff[8] = {-4,-3,-2,-1,1,2,3,4};
             memcpy(edi, edi + voff[(B - 44) >> 2] * stride, width);
             continue;
         }
